+++ history.md
@@ -1,6 +1,12 @@
+# 2022-09-03
+
+- Send "WinClosed" event to node client.
+- Add `onDidFilterStateChange` and `onDidCursorMoved` to `TreeView`.
+
 # 2022-09-02
 
 - Support `diagnostic.virtualTextFormat`.
+- Add command `workspace.writeHeapSnapshot`.
 
 # 2022-09-01